requirements.txt
pyyaml
structlog
pytz
python-dotenv>=1.0.0                # .env 파일 로드 ⭐

# LLM 제공자들
openai>=1.40.0                      # OpenAI GPT
google-generativeai>=0.3.0          # Google Gemini ⭐
requests>=2.31.0                    # Ollama (로컬 LLM)

# HTTP 서버 및 A2A 프로토콜 지원
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.0.0
python-multipart>=0.0.6